# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.2.0",
    "title": "FOUNDLUNTEER API",
    "description":"v1.1.0\n-Add Show saved job for individual\n-Create universal Add, Login, And Get For User\n\nv1.2.0\n-Add get and post image profile for user\n-Register API for Individual\n-Get detail job for Organization"
  },
  "host": "aws.senna-annaba.my.id",
  "basePath": "/",
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "schemes": ["https"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter the token with the Bearer: bearer [token]"
    }
  },
  "paths": {
...
    "/individual/update": {
      "post": {
        "tags": ["Individual"],
        "summary": "Update For Individual",
        "description": "All body request optional",
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "Individual",
          "description": "Add New Individual",
          "schema": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "address": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              },
              "age": {
                "type": "integer"
              },
              "description": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/individual/saveordeletejob": {
      "post": {
        "tags": ["Individual"],
        "summary": "Add or Remove Job Form Individual Job Save List",
        "description": "when the individual wants to save (bookmark) a job based on the job id, the response will be \"saved\", when the individual hits again for the job that has been saved, the job will be deleted from the list and the response will be \"Deleted\"",
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "Job Id",
          "description": "Add Or Remove Job From Bookmark",
          "schema": {
            "type": "object",
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/individual/savedjob": {
      "get": {
        "tags": ["Individual"],
        "summary": "Get Individual Saved Job",
        "produces": ["application/json"],
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Try it out for response"

          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/individual/register": {
      "post": {
        "tags": ["Individual"],
        "summary": "Register Job For Individual",
        "description": "Only individuals with complete data can register.",
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "Job Id",
          "description": "Register Job",
          "schema": {
            "type": "object",
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/individual/registeredjob": {
      "get": {
        "tags": ["Individual"],
        "summary": "Get Individual Registered Job And Process",
        "produces": ["application/json"],
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Try it out for response"

          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/individual/add": {
      "post": {
        "deprecated": true,
        "tags": ["Individual"],
        "summary": "Add Individual",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "Individual",
          "description": "Add New Individual",
          "schema": {
            "type": "object",
            "required": [
              "email",
              "password",
              "name",
              "address",
              "phone",
              "age"
            ],
            "properties": {
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "address": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              },
              "age": {
                "type": "integer"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/individual/login": {
      "post": {
        "deprecated": true,
        "tags": ["Individual"],
        "summary": "Login For Individual",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "Individual",
          "description": "Login For Individual",
          "schema": {
            "type": "object",
            "required": [
              "email",
              "password"
            ],
            "properties": {
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/individual/get": {
      "get": {
        "deprecated": true,
        "tags": ["Individual"],
        "summary": "Get Individual Profile",
        "produces": ["application/json"],
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Try it out for response"

          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}